Migrate JDBC adapter tests to munit
